Change catch_count type from real to integer

+++ R/elementarycatch.R
@@ -17,7 +17,7 @@ elementarycatch <- R6::R6Class(classname = "elementarycatch",
                                  #' @param species_fao_code Object of class {\link[base]{character}} expected. Species code identification on 3 characters.
                                  #' @param species_fate_code Object of class {\link[base]{integer}} expected. Species fate code identification.
                                  #' @param catch_weight Object of class {\link[base]{numeric}} expected. Catch weight in tonnes.
-                                 #' @param catch_count Object of class {\link[base]{numeric}} expected. Catch count in number of fishes.
+                                 #' @param catch_count Object of class {\link[base]{integer}} expected. Catch count in number of fishes.
                                  initialize = function(activity_id,
                                                        elementarycatch_id,
                                                        ocean_code,
@@ -29,7 +29,7 @@ elementarycatch <- R6::R6Class(classname = "elementarycatch",
                                                        species_fate_code,
                                                        catch_weight,
                                                        catch_count) {
-                                   # 1 - Arguments verifications ----
+                                   # 1 - Arguments verification ----
                                    codama::r_type_checking(r_object = activity_id,
                                                            type = "character",
                                                            length = 1L)
@@ -61,7 +61,7 @@ elementarycatch <- R6::R6Class(classname = "elementarycatch",
                                                            type = "numeric",
                                                            length = 1L)
                                    codama::r_type_checking(r_object = catch_count,
-                                                           type = "numeric",
+                                                           type = "integer",
                                                            length = 1L)
                                    # 2 - Attributions ----
                                    private$activity_id <- activity_id
